+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/text/SearchHighlighter.java
@@ -268,8 +268,7 @@ public final class SearchHighlighter {
 	private void bindEvents() {
 		textComponent.getDocument().addDocumentListener((DocumentAdapter) e -> searchAndHighlightResults());
 		textComponent.addPropertyChangeListener(UI_PROPERTY_NAME, new UpdateHightlightColors());
-		selectedSearchTextPosition.when(Objects::nonNull)
-						.accept(this::scrollToPosition);
+		selectedSearchTextPosition.when(Objects::nonNull).consume(this::scrollToPosition);
 	}
 
 	private void scrollToPosition(int position) {